+++ client/server/providers/XmlCompletionProvider.js
@@ -43,9 +43,36 @@ class XmlCompletionHandler extends xmltypes_1.XmlBaseHandler {
                     resolve(this.getElementsInBody(foundCursor));
                 });
             }
+            else if (foundCursor.isInAttribute) {
+                return new Promise((resolve, reject) => {
+                    resolve(this.getCompletionItemsForAttribute(foundCursor));
+                });
+            }
         });
     }
-    getElementsInAttribute(cursor) {
+    getCompletionItemsForAttribute(cursor) {
+        this.logDebug("Processing Tagstring: " + cursor.tagName);
+        let namespace = this.usedNamespaces[cursor.tagNamespace];
+        this.logDebug("Using Namespace: " + namespace);
+        let schema = this.schemastorage[namespace];
+        this.logDebug("Using Schema: " + schema.targetNamespace);
+        let element = this.findElement(cursor.tagName, schema);
+        this.logDebug(() => "Found element: " + element.$.name);
+        let elementType = this.getTypeOf(element);
+        this.logDebug(() => "Found Element type: " + elementType.$.name);
+        let types = this.getBaseTypes(elementType, []);
+        if (types && types.length > 0)
+            elementType.basetype = types[0];
+        let allAttributeTypes = elementType.attribute || [];
+        let base = elementType.basetype;
+        while (base) {
+            allAttributeTypes = allAttributeTypes.concat(base.attribute || []);
+            base = base.basetype;
+        }
+        let matchingAttributeType = allAttributeTypes.find((value, index, obj) => value.$.name === cursor.attribute.name);
+        if (matchingAttributeType) {
+            const attributetype = this.getTypeOf(matchingAttributeType);
+        }
         return undefined;
     }
     getElementsInTag(cursor) {
@@ -56,7 +83,7 @@ class XmlCompletionHandler extends xmltypes_1.XmlBaseHandler {
         this.logDebug("Using Schema: " + schema.targetNamespace);
         let element = this.findElement(cursor.tagName, schema);
         this.logDebug(() => "Found element: " + element.$.name);
-        let elementType = this.getTypeOfElement(element);
+        let elementType = this.getTypeOf(element);
         this.logDebug(() => "Found Element type: " + elementType.$.name);
         let types = this.getBaseTypes(elementType, []);
         if (types && types.length > 0)