+++ ui5xmlserver/src/xmltypes.ts
@@ -132,7 +132,7 @@ export interface FoundCursor extends FoundElementHeader {
 	relativeCursorPosition: number
 	isInElement: boolean
 	isInAttribute: boolean
-	attributeName?: FoundAttribute
+	attribute?: FoundAttribute
 }
 
 export interface FoundElementHeader {
@@ -150,7 +150,7 @@ export interface FoundElementHeader {
 	 * @memberOf FoundElementHeader
 	 */
 	endindex: number
-	elementcontent: string
+	elementHeader: string
 	path: string[]
 	tagName: string
 	tagNamespace: string
@@ -236,7 +236,7 @@ export class XmlBaseHandler extends Log {
 			} else if (cursorPostion > nextMatch.index) {
 				// If the cursor is in the current element
 				return undefined;
-			} else if(!foundElement) {
+			} else if (!foundElement) {
 				return parent;
 			} else if (cursorPostion > foundElement.startindex) {
 				// If it is in the last match part (that means before the current the element)
@@ -262,7 +262,7 @@ export class XmlBaseHandler extends Log {
 			this.logError("Self closing element at root level");
 		// Get first element
 		let parent: FoundElementHeader = {
-			elementcontent: tag[3] ? tag[3] + " " + tag[4] : tag[4],
+			elementHeader: tag[3] ? tag[3] + " " + tag[4] : tag[4],
 			isClosingTag: false,
 			isSelfClosingTag: true,
 			tagName: tag[3],
@@ -323,7 +323,7 @@ export class XmlBaseHandler extends Log {
 				this.logDebug("Found self closing element '" + tag[2] + "'")
 				if (parent) {
 					felement = {
-						elementcontent: tag[3] ? tag[3] + " " + tag[4] : tag[4],
+						elementHeader: tag[3] ? tag[3] + " " + tag[4] : tag[4],
 						isClosingTag: false,
 						isSelfClosingTag: true,
 						tagName: tag[3],
@@ -341,7 +341,7 @@ export class XmlBaseHandler extends Log {
 				}
 			} else {
 				felement = {
-					elementcontent: tag[3] ? tag[3] + " " + tag[4] : tag[4],
+					elementHeader: tag[3] ? tag[3] + " " + tag[4] : tag[4],
 					isClosingTag: false,
 					isSelfClosingTag: false,
 					tagName: tag[3],
@@ -382,34 +382,22 @@ export class XmlBaseHandler extends Log {
 		}
 
 		foundcursor.absoluteCursorPosition = start;
-		foundcursor.relativeCursorPosition = cursorpos;
+		foundcursor.relativeCursorPosition = cursorpos - (foundcursor.tagNamespace.length > 0 ? foundcursor.tagNamespace.length + 1 : 0);
 		foundcursor.isInElement = start > foundcursor.startindex && start <= foundcursor.endindex;
 		foundcursor.isInAttribute = false;
 
 
 		if (foundcursor.isInElement) {
-			foundcursor.isInAttribute = this.textIsInAttribute(foundcursor);
+			for(const attribute of foundcursor.attributes) {
+				if(foundcursor.relativeCursorPosition >= attribute.endpos-attribute.value.length && foundcursor.relativeCursorPosition <= attribute.endpos) {
+					foundcursor.attribute = attribute;
+					foundcursor.isInAttribute = true;
+					break;
+				}
+			}
 		}
 
 		return foundcursor;
-	}
-
-	textIsInAttribute(foundcursor: FoundCursor): boolean {
-		let quote: string = undefined;
-		for (let i = 0; i <= foundcursor.relativeCursorPosition; i++) {
-			switch (foundcursor.elementcontent[i]) {
-				case quote:
-					quote = undefined;
-					continue;
-				case "'": case '"':
-					if (!quote)
-						quote = foundcursor.elementcontent[i]
-					continue;
-				default:
-					continue;
-			}
-		}
-		return quote !== undefined;
 	}
 
 	textGetAttributes(foundElement: FoundElementHeader): FoundAttribute[] {
@@ -424,14 +412,14 @@ export class XmlBaseHandler extends Log {
 		let attributeregex = /\s*?(\w+?)\s*?=\s*?(["'])?/g;
 		attributeregex.lastIndex = foundElement.fullName.length;
 
-		while (amatch = attributeregex.exec(foundElement.elementcontent)) {
-			for (let i = amatch.index + amatch[0].length; i < foundElement.elementcontent.length; i++) {
-				if (foundElement.elementcontent[i] === amatch[2]) {
+		while (amatch = attributeregex.exec(foundElement.elementHeader)) {
+			for (let i = amatch.index + amatch[0].length; i < foundElement.elementHeader.length; i++) {
+				if (foundElement.elementHeader[i] === amatch[2]) {
 					attributes.push({
 						startpos: amatch.index,
 						endpos: i,
 						name: amatch[1],
-						value: foundElement.elementcontent.substring(amatch.index + amatch[0].length, i)
+						value: foundElement.elementHeader.substring(amatch.index + amatch[0].length, i)
 					});
 					attributeregex.lastIndex = i + 1;
 					break;
@@ -492,11 +480,11 @@ export class XmlBaseHandler extends Log {
 					complextype.basetype = basetype;
 				}
 				complextype.schema = schema;
+				complextype.attribute = this.getAttributes(complextype);
 				return complextype;
 			}
 		}
 	}
-
 	protected findElement(name: string, schema: StorageSchema): ElementEx {
 		// Iterate over all
 		for (let element of schema.schema.element) {
@@ -524,7 +512,7 @@ export class XmlBaseHandler extends Log {
 	}
 
 	protected getRightSubElements(element: ElementEx, downpath: string[]): Element[] {
-		let type = this.getTypeOfElement(element);
+		let type = this.getTypeOf(element);
 
 		// Distinguish between sequences and choices, etc. to display only elements that can be placed here.
 		let elements = this.getAllElementsInComplexType(type);
@@ -547,7 +535,7 @@ export class XmlBaseHandler extends Log {
 	}
 
 	/**
-	 * Gets the (complex) type of a given element (with schema)
+	 * Gets the **(complex)** type of a given element (`with schema`)
 	 * 
 	 * @private
 	 * @param {ElementEx} element Element to get the type from
@@ -555,12 +543,13 @@ export class XmlBaseHandler extends Log {
 	 * 
 	 * @memberOf XmlCompletionHandler
 	 */
-	protected getTypeOfElement(element: ElementEx): ComplexTypeEx {
+	protected getTypeOf(element: XmlBase): ComplexTypeEx {
 		try {
 			// Check if complex Type is directly on element
 			if (element.complexType) {
 				let t: ComplexTypeEx = element.complexType[0] as ComplexTypeEx;
 				t.schema = element.ownerschema;
+				t.attribute = this.getAttributes(t);
 				return t;
 				// Check if type is referenced by the element via type="<tname>"
 			} else if (element.$ && element.$.type) {